+++ lib/component/dialog/command.dart
@@ -76,7 +76,7 @@ class _DialogCommandState extends State<DialogCommand> {
                   maxLengthEnforcement: MaxLengthEnforcement.enforced,
                   focusNode: focusNode,
                   controller: controller,
-                  placeholder: widget.hint == null ? null : Text(widget.hint!),
+                  placeholder: widget.hint,
                   // sameSizeHintStyle: true, TODO: fix
                   onChanged: (value) => setState(() {}),
                   onSubmitted: (value) {